setup for easy to add module generator tests
