+++ packages/cdktf-cli/test/get/generator/module-generator.test.ts
@@ -3,7 +3,7 @@ import * as os from 'os';
 import * as path from 'path';
 import { ConstructsMaker, Language } from '../../../lib/get/constructs-maker'
 import { TerraformModuleConstraint  } from '../../../lib/config'
-import { withTempDir } from '../../../lib/util';
+import { expectModuleToMatchSnapshot } from '../util';
 
 test('generate some modules', async () => {
   jest.setTimeout(20000)
@@ -18,25 +18,4 @@ test('generate some modules', async () => {
   expect(output).toMatchSnapshot();
 });
 
-test('no module outputs', async () => {
-  await withTempDir('no-output-module.test', async () => {
-    const curdir = process.cwd();
-    fs.mkdirSync('module');
-    fs.copyFileSync(path.join(__dirname, 'fixtures', 'module-no-outputs.test.fixture.tf'), path.join(curdir, 'module', 'main.tf'));
-
-    const constraint = new TerraformModuleConstraint({
-      source: './module',
-      name: 'module',
-      fqn: 'module'
-    });
-
-    fs.mkdirSync('work');
-    const workdir = path.join(curdir, 'work');
-  
-    const maker = new ConstructsMaker({codeMakerOutput: workdir, targetLanguage: Language.TYPESCRIPT}, [constraint])
-    await maker.generate();
-  
-    const output = fs.readFileSync(path.join(workdir, 'modules/module.ts'), 'utf-8');
-    expect(output).toMatchSnapshot();
-  });
-});
+expectModuleToMatchSnapshot('no module outputs', 'generator', 'module-no-outputs.test.fixture.tf');
