+++ packages/cdktf-cli/test/get/util.ts
@@ -1,8 +1,8 @@
-import { promises as fs } from 'fs';
-import { mkdtemp } from "../../lib/util";
+import * as fs from 'fs';
+import { mkdtemp, withTempDir } from "../../lib/util";
 import { Language, ConstructsMaker } from "../../lib/get/constructs-maker";
 import * as path from 'path';
-import { TerraformDependencyConstraint } from '../../lib/config';
+import { TerraformDependencyConstraint, TerraformModuleConstraint } from '../../lib/config';
 
 export function expectImportMatchSnapshot(constraint: TerraformDependencyConstraint) {
   jest.setTimeout(60_000);
@@ -19,13 +19,38 @@ export function expectImportMatchSnapshot(constraint: TerraformDependencyConstra
 
       await maker.generate()
 
-      const manifest = JSON.parse(await fs.readFile(jsiiPath, 'utf-8'));
+      const manifest = JSON.parse(await fs.promises.readFile(jsiiPath, 'utf-8'));
 
       // patch cdktf version in manifest because it's not stable
       manifest.dependencies.cdktf = '999.999.999';
       manifest.fingerprint = '<fingerprint>';
 
       expect(manifest).toMatchSnapshot();
+    });
+  });
+}
+
+export function expectModuleToMatchSnapshot(testName: string, testCategory: string, fixtureName: string) {
+  test(testName, async () => {
+    await withTempDir(`${testName.replace(/\s*/, '-')}.test`, async () => {
+      const curdir = process.cwd();
+      fs.mkdirSync('module');
+      fs.copyFileSync(path.join(__dirname, testCategory, 'fixtures', fixtureName), path.join(curdir, 'module', 'main.tf'));
+  
+      const constraint = new TerraformModuleConstraint({
+        source: './module',
+        name: 'module',
+        fqn: 'module'
+      });
+  
+      fs.mkdirSync('work');
+      const workdir = path.join(curdir, 'work');
+    
+      const maker = new ConstructsMaker({codeMakerOutput: workdir, targetLanguage: Language.TYPESCRIPT}, [constraint])
+      await maker.generate();
+    
+      const output = fs.readFileSync(path.join(workdir, 'modules/module.ts'), 'utf-8');
+      expect(output).toMatchSnapshot();
     });
   });
 }
